+++ CHANGES.md
@@ -1,4 +1,6 @@
-
+- [csvtk v0.22.1](https://github.com/shenwei356/csvtk/releases/tag/v0.22.1)
+[![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.22.1/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.22.1)
+    - `csvtk pretty`: add header separator line. [#123](https://github.com/shenwei356/csvtk/issues/123)
 - [csvtk v0.22.0](https://github.com/shenwei356/csvtk/releases/tag/v0.22.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.22.0/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.22.0)
     - `csvtk`: